ugtrain: reload mem regions upon lib load in disc stage 1-4
At the moment we read the memory regions only once upon loading of
libmemdisc. But libmemdisc is the first library loaded. Many other
libraries are loaded after it. This is what causes problems finding
the correct memory regions in stage 3 and 4 post-processing.

So read from the "memhack_out" FIFO while waiting for the game
process to end. Reload the memory regions if after two cycles there
is no input from the FIFO anymore. Also read from the FIFO in stage
5 but don't process anything there. Introduce a new structure
"parse_cb" with the callback pointers for read_dynmem_buf() to
reduce its parameters count.