maps: rework readmaps() according to dl_iterate_phdr()
For PIC support we must be able to read from /proc/$pid/maps from
a C library which is preloaded into the game process. The function
dl_iterate_phdr() can't be used as it doesn't update the regions
list when calling it on some systems. But the idea of using a
callback is good.

So rework the whole maps files to use such a callback and separate
C from C++ parts. Make readmaps() a C function again. Also replace
a lot of unclean variable/function names, move getting the exe
path into a separate function and move the maps parsing into
readmaps().