discovery: PIC: add a library backtrace filter
At the moment there is only PIE support for filtering the backtrace
output to code addresses from within the executable. Now that we
have the PIC support in the stage 3/4 post-processing for
unfiltered backtracing it makes sense to filter the backtrace
output to a single library as well.

So parse a further optional section in the discovery string to get
the library name. Then send it as it is to libmemdisc so that it
can do its magic.

Use a discovery string like this:
3[;gbt][;<lib_name>];<size>;0x0;0x0
4[;<lib_name>];<size>;0x0;0x0