libmemdisc: rework exe backtrace filtering with own PIE handling
The backtrace filtering for a library shows that it is not required
to handle PIC/PIE and backtrace filtering externally. It is way too
complex having to call "objdump" to get the start and the end of a
code region. A call to read_maps() also knows about the exe code
region in memory.

So remove the parsing of the start and end backtrace filter
addresses as well as reading PIE info from a FIFO and don't send
PIE info to the FIFO any more. Instead set the backtrace filter
string to ";exe", set the bt_filter to "\0", call read_maps()
and handle PIE in its callback function set_bt_addrs() like we
handle PIC filter support. This simplifies the whole thing a lot.
By default everything becomes unfiltered.

New unfiltered discovery strings:
3[;gbt];<size>
4;<size>

New discovery strings with exe filter:
3[;gbt];exe[;<heap_saddr;heap_eaddr];<size>[;<code_addr>]
4;exe[;<heap_saddr;heap_eaddr];<size>[;<code_addr>]