libmemdisc: handle PIC externally
On high security systems we've noticed that the game process isn't
always allowed to read from /proc/$pid/maps itself. So let ugtrain
do that and send the library code start and code end through the
FIFO.

But attention: It is possible that the same library is loaded
multiple times. So store a backup of the code address for backtrace
filtering before adding the code start of the library to it. Also
delay the activation until ugtrain sends a message that it is ready
to handle PIC. Otherwise we might miss library load messages.
Search for the library in the maps before sending the notification
in form of a library backtrace filter. The library might be loaded
early and doesn't appear in dlopen() this way. Send unfiltered if
it can't be found.

Tested against btanks 0.9.8083-4 on Ubuntu 14.04 x86_64 with the
missiles of the own tank (starting at 15) with discovery commands:

$ ugtrain -D "2" -S btanks
$ ugtrain -D "4;1008" -S btanks
$ ugtrain -D "4;exe;1008" -S btanks
$ ugtrain -D "4;libbt_objects.so;1008" btanks
$ ugtrain -D "4;libbt_objects.so;1008" -S btanks
$ ugtrain -D "4;libbt_objects.so;1008;0x36b75" -S btanks
$ ugtrain -D "3;gbt;libbt_objects.so;1008" -S btanks
$ ugtrain -D "3;gbt;libbt_objects.so;1008;0x36b75" -S btanks

The libbt_objects.so gets loaded twice with _Znwm() at 0x36b75.

Early library load has been tested with chromium-bsu 0.9.15.1-1:

$ ugtrain -D "4;libstdc++.so.6.0.19;288" chromium-bsu64
$ ugtrain -D "4;libstdc++.so.6.0.19;288" -S chromium-bsu64