doc: dynmem: update hacking and discovery to current state
There is still the libmemhack config in static memory mentioned
but everything is stored on the heap now. So update that. The PIC
support for the discovery has changed a lot. So document the new
behavior. Use PIE for the Chromium B.S.U. example as PIE is the
default now on most Linux systems.