cfgparser: enhance PIC lib parsing for grow
Parsing a library name for the dynamic memory allocation entry
isn't enough. We also need it for growing objects. There is also
absolutely no check for the name of the library.

So put everything into the inline function parse_pic_lib() and call
it for dynmem and grow entries. Also check the library name for
blanks and the word "exe". Assume that the executable is meant by
that.