discovery: add process_discovery() to move code to here
There is still too much discovery code between the other ugtrain
code. So put all that into a new function process_discovery().
This also gives us the chance to get rid of the variable call_pid.
With wait_orphan() and proc_to_pid() in run_game() it is not
required anymore.