ugtrain: update copyright for 2015 in all src files
Put the copyright into a single line. This way feature contributors
can be named below and it can be updated by scripts easier. Set
everything to upto 2015 as work goes on and on no matter what. This
thing is too elite to be thrown to the graveyard of discontinued
FOSS projects! The fight continues attracting users and contributors!