libmemhack: transfer PIC library names to here
In order to support PIC it is required that ugtrain sends the names
of the libraries from where allocations are done to libmemhack. So
send them, receive them and output them for now. Also replace ioffs
handling with moving a parsing start pointer.

During testing a bug in the dynmemgrow handling has been found
resulting duplicate config entries to be sent. So enhance the
checks to fix this issue.