libmemhack: handle early PIC and PIC/PIE with growing
Games which use memory allocations from within a library (PIC) are
not supported yet but we urgently need this support in order to
support more games. There is also no PIC/PIE handling for growing
memory objects. The PIE handling is even limited to only 5 classes.

So rework that and let ugtrain find the code offsets of early
loaded libraries. Send exe and early PIC code offsets to libmemhack
for every class and also for the growing. In libmemhack parse one
by one and add the transferred code offsets to their respective
code addresses in the config. Output the new config afterwards.
Also add the code offsets in ugtrain to the code addresses and
don't use the single global code offset from options any more.