ugtrain: introduce basic testing support
At the moment we have no possibility to stop processing at a certain
point for testing the functionality of certain modules like the
options parser, the config parser or the config output. But this
would be essential to write automated test cases.

So introduce the configure option '--enable-testing' which injects
an application option '--exit-after=<part>' where the possible parts
are 'optparsing', 'cfgparsing' or 'cfgoutput'. Use inline functions
in order to check for this option and to exit at the related code
location.