adapt: common: add pr_err and pr_dbg to write dbg output to stderr
There is a common pattern in the debug output always requiring to
check if DEBUG is set. So move that into a new function pr_dbg()
to make the code more readable. Write to stderr to separate it from
the actual adaption output. Also introduce pr_err() to write an
error message with a prefix to stderr.