ugtrain: support spaces in the game path
The issue with spaces in the game path is that we split a command
string into a command vector at the locations of spaces.
Furthermore, execvp() doesn't support any spaces in the command
file name - not as is, not with "quotes" and not with inserting a
backslash in front of the spaces.

So run commands with spaces in the game path with the help of the
shell, place a backslash in front of each space in the path and
don't split the command vector in that case.