ugtrain: always run the game in a shell
The ugtrain can handle all kinds of loaders which fork or exec by
now. So it works running the game in a shell. The execvp() function
can't handle spaces. Also maintaining two code paths for running
with the shell and without it is too complex in the code. So always
run the game in a shell and remove cmd_str_to_cmd_vec(). Running
the game gets slower but code becomes simpler and more flexible.