adaption: adapt 'proc_name' also without 'game_binpath' set
Currently the proc_name can only be adapted if also game_binpath is
set in the config. Also the game_call has to be different from the
proc_name. This is kind of limiting.

So skip parsing of the game_binpath in the adaption result if it's
not set in the config, set the new game_call if it was identical
with the proc_name and determine game_path and game_binpath again
after adaption. Exit only there if complete paths can't be
determined.