adapt: common: allow disabling the explines check
With small object sizes it is really difficult to find unique code
calls but it might be possible that we just need the very first
one. In that cases we don't really care for what's coming after
this. So disable the explines check if this variable is set to 0 or
a negative value.