adapt: common: use negative explines as selected head lines
There are cases where we don't care how many lines there are in
total but instead we only want to cut at a certain location and
process the first part. Supertux 2 is an example for that as the
malloc size is small and only the first _Znwm/_Znwj call is the
right one. So cut at the location given by negative explines.