fifoparser: simplify input handling and parsing
Doing frequent calculations is cumbersome and error-prone. This
especially affects the input and parsing positions (ipos and ppos).
So maintain an input and a parsing start pointer (istart and pstart)
instead. Precalculate the maximum number of bytes to be read and
reduce code duplication by restructuring. Also remove debug messages.
With simpler code it is easier to review and these aren't required
any more.

Functionality remains almost completely the same.