discovery: fix 100% CPU load in stage 1/2/3/4 at game exit
The worker executing run_stage1234_loop() consumes 100% CPU load
when the game process exits before scanmem. This has been noticed
while running the Chromium B.S.U. example like this:

$ ugtrain -D 2 -S chromium-bsu64

The reason for this is that sleep_sec_unless_input() returns
immediately and also the non-blocking read returns immediately. So
check if the game is still running if nothing is read and sleep one
second mandatorily if it's not.