fifoparser: split the parsing from the FIFO reading
We have a problem with uninitialized dynamic memory objects. The
idea is to queue the mallocs in a buffer and to process them later.
So introduce parse_dynmem_buf() and move the parsing code to that
function. Also add the message end to the parsing parameters so
that we can extract the message and write it to a buffer.