ugtrain: delay malloc processing by one cycle
We have a problem with uninitialized dynamic memory objects. We
plain process them too fast (in the same moment as the frees). So
queue them for one cycle and delay the processing by further 250ms
so that the game has a chance to initialize the memory objects in
the meanwhile.

This has been stress tested by building a lot of droids in Warzone
2100 fast until the droid maximum has been reached. No valid object
has been kicked out. Also with Chromium B.S.U. no uninitialized data
is visible any more. Delaying by 100ms wasn't enough there.

Fixes: GitHub issue #7