memmgmt: remove mallocs to be freed from the malloc queue
There is the unlikely event that a malloc has been put to the malloc
queue and before processing the queue a free occurs for that malloc.
This means processing of a temporary object. If we wouldn't remove
that malloc from the queue, then the game might crash with SIGSEGV.
So introduce the function remove_malloc_from_queue() to do exactly
that.