treewide: remove Open Game Cheating Association mentions
The whole idea behind the Open Game Cheating Association is a fail.
Can't unite the open game cheating developers or users, can't get
enough users/testers for ugtrain. Not enough people watch or like
the videos. This tool gets abused right away and not enough people
are willing to help in any way. No regular contributions. There is
not enough interest in the whole topic. This is also why scanmem/
GameConqueror had to be resurrected.
So remove all mentions of the Open Game Cheating Association.

Fixes: GitHub issue #12