lib: move C list processing (list.h) to here
As described in GitHub issue #8 the hardened build fails due to
C/C++ mix in the lib. This means we need to replace C++ list
processing with C list processing for the memory regions. So move
list.h to the lib to prepare that and also move offsetof() and
container_of() to this file as we don't need it anywhere else.