ugtrain: replace C++ with C list processing for rlist
As described in GitHub issue #8 the hardened build fails due to
C/C++ mix in the lib. So replace the C++ list processing with C list
processing for the memory regions (rlist).

This has been tested with the game Battle Tanks (btanks) as it uses
late PIC handling (reloads the rlist upon dlopen()) and with the
default game Chromium B.S.U. - works as expected.

Reported-by: @ignatenkobrain
Fixes: GitHub issue #8