libmemhack: convert the mutex into rwlock
The arrays of stored malloc() addresses to be processed by free()s
are often read and are read by many threads. Updates are very rare
(only done by configured mallocs). So it doesn't make much sense
to limit them to a single thread accessing the data structures at
a time with the regular mutex. The libpthread doesn't only provide
a mutex lock but also rwlocks which allow having multiple readers
at a time.

So switch over to the rwlock for performance reasons. Protect
malloc()s with the write lock and free()s with the read lock.

This has been tested with Warzone 2100 in a 7 bots against 1 player
game killing 530 enemy droids.

Fixes: GitHub issue #24