examples: bloodfrontier: add checks and rifle ammo
Currently also the enemy bots get unlimited health as FPSent1
objects are also created for them. So add an object check if we are
the local player discovered with ugsimfind and add checks for
everything else to make the config more reliable. Also add the
locking of the rifle ammo.