examples: add a Mini-Sub V1.0 config for Raspbian/Raspberry Pi
The Raspberry Pi uses the ARM architecture without any security
module. It is easy to clone the scanmem and the ugtrain source
there and to compile directly on that platform. This is a good
chance to demo that ugtrain works on ARM as well. Most games are
crap or running too slow there. The game Mini-Sub is an exception.
It is fast, simple and leaves room to find the interesting memory
values.

So add a +2 config for Mini-Sub on Raspbian locking lives and ammo.