libmemdisc: fix a stage 0 segfault by calling atexit() later
Currently the atexit() function is called and therefore registers
the final flush too early. It is called before parsing the data on
the input FIFO. If something goes wrong during parsing or an
invalid stage is selected, then all files are closed by do_cleanup()
but the final flush is still registered. This causes a segfault when
ending the game.

So register the final flush after input parsing when nothing can go
wrong any more.