discovery: limit printing of matching mallocs to 15
The process of reverse-matching the heap address from memory
scanning to the recorded mallocs is unlimited. But sometimes the
interesting memory object is allocated very late and a lot of
mallocs are matching. We are usually only interested in the first
matching malloc unless we did a mistake and e.g. returned to main
menu causing further matching mallocs. But this is usually still
detectable within the first 15 matches. Printing too many matches
can be really confusing. So limit printing to 15 matches.