ugtrain: ensure always using the right game PID
With pure static memory configs for games which use a loader, the
ugtrain run_game() function returns the wrong PID (the one of the
loader) causing ugtrain to fail to attach to the correct PID. The
order of the process state check and the pid check is incorrect.

So switch the order and ensure that a valid game PID (which doesn't
belong to a game instance which ran before) is determined first.
Only with a valid new game PID check if the process belonging to
it is also running.

Example: freedoom1 shell script calls boom, ammo in static memory

Fixes: GitHub issue #29