ugtrain: enhance game start error and timeout handling
If the game start times out, then there is no further check which
causes run_game() to return regularly instead of jumping into the
error case. It is possible to attach to a zombie process this way.
Also the pid isn't checked again. Timeouts use magic numbers and
the delay before scanmem attaches is with 3s lower than the game
start timeout (5s).

So add the missing checks, use defines for the numbers and reduce
the game start timeout to 2.8 seconds.