system: read the process name from /proc/$pid/status
The game freedoom1 shows that the process name is not reliably
equal to the file name from the /proc/$pid/exe. The freedoom1 shell
script calls 'boom', the process is called 'boom' but the exe
symlink points to 'prboom-plus'. So read the first line from
/proc/$pid/status to check the process name in check_process().

This has been tested against all PC examples.