examples: add a freedoom1 64 bit config
Freedoom1 is quite popular as the original Doom was also very
popular. It was one of the first first-person shooters and it was
completely in 2D.
Freedoom uses the FLOSS PrBoom-Plus engine. The new bit here is
that the /proc/$pid/exe symlink is not equal to the process name.
The symlink points to 'prboom-plus' but the process name is 'boom'.
Most interesting values are stored in static memory without PIE.
Like always with FPS games, health and ammo are the interesting
values. So add a +2 trainer config for that.

This game should also be used as an example to demo ugtrain within
a tutorial video as the game is popular, data structures are simple
and with 2D graphics it can be recorded together with a terminal.