linuxhooking: move malloc hooks to a separate file
The libraries libmemdisc and libmemhack use the same code to hook
on memory allocation functions. It is hard to maintain it twice
and code duplication is always bad. Gcc doesn't inline the malloc
post-processing function any way and we haven't noticed any
performance issues caused by ugtrain so far.

So move the malloc hooks to a separate file 'memhooks.c'. Let this
file use the interfaces postprocess_malloc(), preprocess_free(),
and postprocess_dlopen().