libprivacy: just block network sockets
Libprivacy is not really the right thing to configure which network
connections should be allowed. It is mainly meant against abuse by
beginners who don't know what they do. So change the code back to
just hooking socket() and socketpair() to block everything which is
not AF_LOCAL. That should mainly enforce single player cheating if
dynamic memory cheating is involved. It is most important that
nothing revealing ugtrain is sent to publishers of commercial games.