ugtrain: Init ~/.ugtrain/ with the examples upon first run
When installing the Debian package, the user might want to try out
the examples. Also testing adaption on different Linux live-USBs
requires the examples to be in place. So check for ~/.ugtrain/ when
starting ugtrain. If it doesn't exist, create it. Copy the examples
from /usr/share/doc/ugtrain/examples/ to it and extract them with
'gunzip'. Show the user what's going on to become aware of the user
config directory.