configure: Skip Linux-specific stuff for Windows builds
Running the configure command with MinGW (--host=i686-w64-mingw32)
fails. It tries to detect pthreads which are not available on
Windows. Also other features are affected.

So detect Windows by checking for windows.h earlier and disable
all features which don't apply to Windows. These are: pthreads,
glib, multilib and bash completion.