cfgentry: Add code_offs to dynmem and grow
The game Battle Tanks (btanks) is one of the rare games which comes
with game object allocation from within a library (libbt_objects.so).
It is even more rare to load that library with dlopen() while the
game is already running.
Late PIC handling is required to support that. Ugtrain needs to
store a copy of the configured code address as a code offset. So
load addresses are never added to it. This prevents that early and
late PIC handling collide and also covers loading the library
multiple times. The game btanks loads libbt_objects.so twice.

So add a code_offs member variable to the classes DynMemEntry and
GrowEntry.