fifoparser: Move 'struct mqueue' declaration to here
The structures used as void pointer argp for read_dynmem_buf() and
parse_dynmem_buf() should be also located in the FIFO parser header
file. The parsing covers aspects of multiple code parts. Only at
this central location it is possible to unite them.
So move the mqueue declaraton from memory management to the FIFO
parser.

This is a preparation for memory hacking with late PIC support.