fifoparser: Add 'struct lf_params' for late PIC handling
Late PIC handling for discovery is already supported and has its
own library processing function (lf) call-back parameters. For
dynamic memory hacking, different parameters are required. It also
needs the game PID and the output file descriptor (ofd) but it
requires the config entry list (cfg) and the regions list (rlist).
With that information the regions list can be reloaded and the
configured code addresses for mallocs can be updated.

So declare struct lf_params with that data.