fifoparser: Add 'struct dynmem_params' for late PIC handling
It is required to pass mf(), ff() and lf() parameters to the FIFO
parser. mf() and ff() use struct mqueue and the lf() function
requires struct lf_params. So put pointers for both into a new
struct dynmem_params.

This is a preparation for memory hacking with late PIC support.