cfgparser: Initialize the code_offs members
The code offset member variables for the classes DynMemEntry and
GrowEntry have been introduced for late PIC memory hacking handling.
The difference to the code address members is that no load addresses
are added to them. So initialize them with the code address member
value.

This is a preparation for memory hacking with late PIC support.