libmemhack: Fill postprocess_dlopen() with late PIC support
When a library is loaded by the game process at runtime, then
libmemhack has to extract the library name, check if that name is
configured, get the new load address of that library, and set the
new code addresses for it.

So implement that functionality into postprocess_dlopen() for late
PIC support and also implement that for grow handling. Only keep
the communication with ugtrain as a stub by introducing the function
get_lib_load_addr(). Cache the load address to only communicate once
with ugtrain and print the new load address only once.