libmemhack: Introduce send_lib_name()
The library name gathered by hooking dlopen() has to be sent to
ugtrain so that libmemhack can receive the library load address.
That is required for late PIC support.

So introduce the function send_lib_name() which just sends the
library name through the FIFO. Let it return 0 upon success and -1
in case of an error. Use a 4096 bytes buffer, snprintf() and the
write() function as always.