libmemhack: Fill get_lib_load_addr() with code
Now that send_lib_name() is in place, send the library name from
dlopen() hooking to ugtrain and read back the library load address.
Parse it and only take it over if that succeeds. Return 0 for
success and -1 in error case.

This completes late PIC support and has been tested with the game
Battle Tanks (btanks) as it loads libbt_objects.so with dlopen() at
runtime. The game loads that library twice and allocates all
interesting game objects from it.