examples: Add a Battle Tanks +3 config to demo PIC
The game Battle Tanks (btanks) allocates everything from libraries.
This is why it is the perfect example to demo the PIC handling of
ugtrain. libbtanks_engine.so allocates the memory object with the
money value and is loaded early. This means it is visible with the
"ldd" Linux command and is loaded before running main(). But most
interesting stuff is stored in libbt_objects.so. The vehicle objects
Launcher, Shilka and Tank containing the health value are allocated
from that library. Also the objects containing the ammo for the
weapons are allocated from there. The game is a strategy game where
the player chooses one of these three vehicles and fights against
others. For shooting at them, the player gets money.

So add a Battle Tanks +3 config locking money, health and ammo.