doc: statmem: Update to current state
Some things changed and some things are not completely clear. So
rework the documentation a bit to the current state.

It is outdated that ugtrain would attach to the game instance which
has been started first. It is the most recently started one instead.
Also describe the issue about ptrace() security a bit more with a
hint to the Yama security module.