maps: Move code to open maps to fopen_maps()
It is required to dump the maps file to a given file when dumping
all dynamic memory objects. For not having to duplicate code, just
move the common code to open maps with fopen() to an inline function
called fopen_maps(). Construct the file path from the given PID and
set the given file pointer. Also return a status if fopen() succeeds.

Reference: GitHub issue #14