maps: Introduce write_maps_to_file()
A function is required to create a copy of the /proc/$pid/maps file.
This can help when dumping dynamic memory objects as these can
contain pointers. With the maps information it is possible to check
to which memory region these are pointing to. So introduce the
function write_maps_to_file() for this. Use fopen(), fread() and
fwrite().

Reference: GitHub issue #14