dump: Introduce dump_maps() to dump game memory maps
The function write_maps_to_file() has been introduced and can write
a copy of the game memory maps file to any location. It is useful
to find the memory regions for pointers in dumped memory objects.
So introduce the function dump_maps() to use it to dump the maps to
"maps.dump.txt". If the file already exists, create a backup named
"maps~.dump.txt" first to always keep two states.

Reference: GitHub issue #14