memhooks: Introduce a function hooking template
There is too much code duplication in the memory hooks. It is better
maintainable to handle that with a macro template. Then things can
be changed at a central location. Also dlsym() is called at two
locations. So introduce a macro for dlsym(), the function hooking
template __HOOK_FUNCTION() and further helper macros to replace
the malloc() hook for now.

Only three code locations differ between the hooks: no_hook code,
pre-call code, and post-call code. The no_hook part is usually just
the dlsym() call. The only exception is calloc(). So introduce a
macro _HOOK_FUNCTION() to set the default no_hook code. The pre-call
code can be used for free() and the post-call code can be used for
malloc(). Other parameters are the function type, name, name as
string, parameters, parameters without types, local variables and
the variable to be returned.