memhooks: Hook _Znwm() and _Znwj()
The reverse stack offset from malloc() back to _Znwm() differs from
distribution to distribution and adapting it is really cumbersome.
We need to get closer to the call within the game binary and to
lower the reverse stack offset so that it is always the same value.

So hook _Znwm() for 64-bit systems and _Znwj() for 32-bit systems
as this is the C++ "new" operator. The disadvantage is that many
examples need to be changed for this.

Reference: GitHub issue #40