adaption: Drop stack offset adaption
The goal must always be having a static reverse stack offset as low
as possible so that it remains static. It should always be 0x8 on
x86_64 and 0x4 on x86/i686. This way it is not required to adapt it.
When we've adapted it in the past, then the offset for e.g. _Znwm()
was always the same for the whole platform. It doesn't make sense
to adapt it on a per-game basis. So drop the reverse stack offset
adaption.