lib: Compile memattach functions inline
The only time-critical ugtrain function visible in the Valgrind
callgrind output during dynamic memory hacking is memread(). The
memattach functions are usually small and/or time-critical. So
complile them all inline.

There is no performance bottleneck and the difference can't be
measured as this is called rarely. But arbitrary configs can be
put together. So just be sure.