lib: Fix compiler warnings in memattach functions
Inlining the memattach functions revealed compiler warnings. The
variable 'pos' is an off_t and is compared against size_t resulting
in comparing signed and unsigned. So change this variable to size_t.
Also calculations with a void pointer causes a warning. So cast the
variable 'buf' from 'void *' to 'char *' internally.

The functionality remains the same.