discovery: Add stage 4 beginner support
A beginner shouldn't have to know about the discovery stages. He
just enters the memory size and then stage 4 should be used
automatically. So send the size provided by the beginner directly
to libmemdisc but check if it is greater than or equal to 8 and if
only digits have been provided.