ugtrain: Add pointer following support from static memory
If adaption is not required or LD_PRELOAD is not available, then it
might be beneficial to support single dynamic memory objects by
pointer following from static memory. It can also help to debug
issues in other methods. So add this feature. But that method
remains unpreferred as memory discovery is likely cumbersome and
requires a debugger.