examples: Add Chromium B.S.U. pointer following
Pointer following is the method of choice with Cheat Engine on
Windows. Ugtrain also supports to follow a pointer in static memory
to a single dynamic memory object on the heap. So demo that to show
the difference to the preferred API hooking method. There is no easy
way to adapt this and it is also hard to discover all interesting
values within the memory object.