cfgparser: Add hex value support
E.g. for checking pointers it is beneficial to support hex values.
The problem is that values starting with "0x" are already used for
value at address calls. So use the prefix "va0x" for these and
detect hex values by the "0x" prefix. Set the number base to 16 for
parsing the numbers and check them with isxdigit() if the number
base is 16.

Attention: Existing configs need to be adapted for this.