maps: Don't stop parsing file paths at spaces
The parsing of 'map.file_path' is done with the sscanf() '%s' format
specifier. This stops parsing at spaces resulting in incomplete file
paths. This also affects the region type and load address detection.
So use the format specifier '%[^\n]' instead as it matches everything
but newline characters.

This is a port of a scanmem commit.