fifoparser: Handle negative tmp_ilen correctly
There is a check which should prevent negative tmp_ilen to be passed
to memset(). It has been extended by a further condition so that
there is a case that it can still be passed as negative number.

if (tmp_ilen <= 0 && argp)
        return -1;

This has been found by Coverity. So set negative numbers to 0 before
calling memset().

if (tmp_ilen <= 0) {
        tmp_ilen = 0;
        if (argp)
                return -1;
}

Reference: GitHub issue #42