ugtrain: Add a NULL check to make Coverity happy
In the hot game trainer path when processing an object check,
'dynmem' is expected to be valid as the parser has to ensure that
but Coverity still thinks there is a defect. So only dereference
'dynmem' if it is not NULL but make the check for 'dynmem' likely
for optimization.

Reference: GitHub issue #42