dump: Store a backup of pointer memory
For dynamic memory it is already possible to compare two states.
That should be possible for pointer memory as well. So store a
backup if the target file for the dump already exists.

Example: rename p0_000.dump to p0_000~.dump

This is useful for vector discovery within a dynamic memory object.
C++ vectors are stored as a start and an end pointer in memory. This
way we can use pointer following and know the size of the vector.