tools: Introduce ugptrfind
Ugptrfind is used to locate C++ vectors and pointers to a specified
memory region (usually the heap) in a dynamic memory object dump
file. It is limited to aligned search for now and doesn't get the
region start and end pointer automatically. This information can
be gathered manually from the dumped maps file.

Example: ugptrfind 0_000.dump 0x01ebc000 0x0bc9b000 | less

4c0: 0x8034c80
4c8: 0x8035150 <-- vector 1232 bytes

This is the struct stored_torpedo vector of the own submarine object
in Danger from the Deep. The vector size is determined from end
pointer - start pointer.