ugtrain: Introduce a wish value per memory object
Using the same wish value for all dynamic memory objects of the
same class can cause the game to become unstable. That is especially
the case for dynamic wish values like "min" or "max".
An example is the ship class in the game endless-sky. Different
player ship types have different maximum shields values.

So introduce a vector for the wish values per object.