README: Rework for 2017
Mark that this is a research project, that I don't have the time
to develop a GUI, don't need one, and focus on researching game
cheating methods on the cost of usability instead.

I'm not getting any contributions. So don't ask for that any more.
I'll rather focus on my personal needs now.

List distribution packages and update current limitations.