discovery: Add value on stack support
Add a function to get the stack start from /proc/$pid/maps and to
store it in opt. Call that one in stage 1,2,3,4 discovery. Add an
sf() callback for the FIFO parser which gets the stack end from
libmemdisc, stores it in opt, and matches the provided memory
address from scanmem to the stack area. Also pass opt to the sf()
callback for libmemhack and optimize that function.

Due to reverse matching of mallocs first, it takes a bit until the
stack end is reached in the discovery process. Looks like this:

$ ugtrain -D -S endless-sky
...
1> list
[ 0] 7ffc2e8acfa0, 36 +        1dfa0, stack, 49000, [I64 I32 I16u ]
1> exit
Memory address (e.g. 0xdeadbeef): 7ffc2e8acfa0
Searching reverse for 0x7ffc2e8acfa0 in discovery output..
stack contains 0x7ffc2e8acfa0, offs: 0x4a0

The number of credits in the game endless-sky has been found here
as forward stack offset 0x4a0 which can be put to the endless-sky
config.

Reference: #44