ugtrain: Add pointer to C string watching support
Sometimes it is only possible to locate a dynamic memory object by
searching for its (unique) name and then searching for the found
char pointer. That was the case for the space ships in endless-sky.
It would be nice to be able to watch that string then to better
identify the object we are looking at with ugtrain.

So add support for watching C strings with upto 128 chars from a
pointer. The string has to be copied from game memory to ugtrain
memory. So introduce a new 'cstr' member to CfgEntry and a respective
vector. Print the value the pointer has and the string when following
that pointer. This has been tested with endless-sky for dynamic memory,
with a small test program for static string in static memory, but
not tested for pointer memory so far. A mechanism to settle pointers
might be required for static memory in future.

Config example:
dynmemstart Ship 2808 0x519295 0x8
Name 0x98 cstrp watch
dynmemend

Output example:
*Ship[1] = 0x24651f0, 6 obj.
Name[1] at 0x2465288, Data: 0x1f58d68 ("Diamond Grey")

This is a captured pirate ship in the game endless-sky.