examples: Add Endless Sky config
The game endless-sky is a perfect example for the new value on
stack support as it stores the number of credits as i64 there. The
configured stack value forward offset is subtracted from
__libc_stack_end provided by libmemhack to get the virtual memory
address of the current game run reliably.

It is also a bit unusual that the game stores the visible attributes
of the own ships as doubles. I could only find my own ship in memory
by searching for its unique name and searching for the found unique
i64 pointer again.

This is why the C/C++ string watching support by pointer following
has been introduced. This way ugtrain can show the name of the ship
in its output.