ugtrain: Make value from reference more reliable
Config entries which get their wish value from another config entry
need to be synchronized. Getting only the old value means that the
config entry with the reference might not get the value which has
been written to target memory in the same cycle. So remember and
check that state and use the wish value instead if that state is
set.

Another problem is pointer following. In warzone2100 v3.2.x, setting
SBody to SMaxBody causes that the value from the unsettled pointer
is used and SBody is increased one cycle before SMaxBody. The game
might detect that something is wrong. So wait until the pointer is
settled, return value 0 before and handle value 0 by skipping the
writing to target memory.